requirements.txt
ChatterBot==0.5.5
defusedxml==0.4.1
Django==1.10.5
djangorestframework==3.5.3
freeze==1.0.10
future==0.16.0
jsondatabase==0.1.7
nltk==3.2.2
oauthlib==2.0.1
PyJWT==1.4.2
pymongo==3.4.0
python3-openid==3.0.10
python-social-auth==0.3.5
python-twitter==3.2.1
requests==2.13.0
requests-oauthlib==0.7.0
six==1.10.0
social-auth-core==1.0.1
virtualenv==15.1.0
whitenoise==3.3.0
